fix: move utils around and the config
