+++ src/fides/cli/commands/db.py
@@ -5,7 +5,7 @@ from fides.cli.options import yes_flag
 from fides.cli.utils import with_analytics
 from fides.common.utils import handle_cli_response
 from fides.core import api as _api
-from fides.core.utils import echo_red
+from fides.common.utils import echo_red
 
 
 @click.group(name="db")
