+++ src/fides/core/push.py
@@ -9,7 +9,7 @@ from fideslang import FidesModel, Taxonomy
 from fides.common.utils import handle_cli_response
 from fides.core import api
 from fides.core.api_helpers import get_server_resources
-from fides.core.utils import echo_green, echo_red
+from fides.common.utils import echo_green, echo_red
 
 
 def sort_create_update(
