+++ src/fides/core/system.py
@@ -13,7 +13,7 @@ from fides.core import api
 from fides.core.api_helpers import get_server_resource, get_server_resources
 from fides.core.filters import filter_aws_systems
 from fides.core.parse import parse
-from fides.core.utils import echo_green, echo_red
+from fides.common.utils import echo_green, echo_red
 
 
 def generate_redshift_systems(
